progress in moving Codec to 84
