+++ src/main/java/com/amazon/opendistroforelasticsearch/knn/index/codec/KNN80Codec/KNN80Codec.java
@@ -51,7 +51,7 @@ public final class KNN80Codec extends Codec {
     private final CompoundFormat compoundFormat;
     private Codec lucene80Codec;
 
-    public static final String CODEC_NAME = "KNN80Codec";
+    public static final String KNN_80_CODEC_NAME = "KNN80Codec";
     public static final String LUCENE_CODEC = "Lucene80"; // Lucene Codec to be used
     // Lucene version for the Codecs Doc Value Format. Note that this is not always the same as LUCENE_CODEC. Sometimes
     // a Codec version will use an earlier Codec version's Doc Value Format. For instance Lucene 84 uses Lucene 80
@@ -59,7 +59,7 @@ public final class KNN80Codec extends Codec {
     public static final String LUCENE_DOC_VALUES_FORMAT = "Lucene80";
 
     public KNN80Codec() {
-        super(CODEC_NAME);
+        super(KNN_80_CODEC_NAME);
         this.docValuesFormat = new KNN80DocValuesFormat();
         this.perFieldDocValuesFormat = new PerFieldDocValuesFormat() {
             @Override
@@ -133,23 +133,5 @@ public final class KNN80Codec extends Codec {
     @Override
     public PointsFormat pointsFormat() {
         return getDelegatee().pointsFormat();
-    }
-
-    public static KNNCodecUtil.Pair getFloats(BinaryDocValues values) throws IOException {
-        ArrayList<float[]> vectorList = new ArrayList<>();
-        ArrayList<Integer> docIdList = new ArrayList<>();
-        for (int doc = values.nextDoc(); doc != DocIdSetIterator.NO_MORE_DOCS; doc = values.nextDoc()) {
-            byte[] value = values.binaryValue().bytes;
-
-            try (ByteArrayInputStream byteStream = new ByteArrayInputStream(value);
-                 ObjectInputStream objectStream = new ObjectInputStream(byteStream)) {
-                float[] vector = (float[]) objectStream.readObject();
-                vectorList.add(vector);
-            } catch (ClassNotFoundException e) {
-                throw new RuntimeException(e);
-            }
-            docIdList.add(doc);
-        }
-        return new KNNCodecUtil.Pair(docIdList.stream().mapToInt(Integer::intValue).toArray(), vectorList.toArray(new float[][]{}));
     }
 }
