+++ src/main/java/com/amazon/opendistroforelasticsearch/knn/index/codec/KNN80Codec/KNN80DocValuesConsumer.java
@@ -85,7 +85,7 @@ class KNN80DocValuesConsumer extends DocValuesConsumer implements Closeable {
             String indexPath = Paths.get(((FSDirectory) (FilterDirectory.unwrap(state.directory))).getDirectory().toString(),
                     hnswFileName).toString();
 
-            KNNCodecUtil.Pair pair = KNN80Codec.getFloats(values);
+            KNNCodecUtil.Pair pair = KNNCodecUtil.getFloats(values);
             if (pair == null || pair.vectors.length == 0 || pair.docs.length == 0) {
                 logger.info("Skipping hnsw index creation as there are no vectors or docs in the documents");
                 return;
