--- src/main/java/com/amazon/opendistroforelasticsearch/knn/index/codec/KNN84Codec/KNN84DocValuesReader.java
@@ -1,58 +1,0 @@
-/*
- *   Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- *   Licensed under the Apache License, Version 2.0 (the "License").
- *   You may not use this file except in compliance with the License.
- *   A copy of the License is located at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   or in the "license" file accompanying this file. This file is distributed
- *   on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- *   express or implied. See the License for the specific language governing
- *   permissions and limitations under the License.
- */
-
-package com.amazon.opendistroforelasticsearch.knn.index.codec.KNN84Codec;
-
-import com.amazon.opendistroforelasticsearch.knn.index.codec.BinaryDocValuesSub;
-import org.apache.lucene.codecs.DocValuesProducer;
-import org.apache.lucene.index.BinaryDocValues;
-import org.apache.lucene.index.DocIDMerger;
-import org.apache.lucene.index.EmptyDocValuesProducer;
-import org.apache.lucene.index.FieldInfo;
-import org.apache.lucene.index.MergeState;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * Reader for KNNDocValues from the segments
- */
-class KNN84DocValuesReader extends EmptyDocValuesProducer {
-
-    private MergeState mergeState;
-
-    KNN84DocValuesReader(MergeState mergeState) {
-        this.mergeState = mergeState;
-    }
-
-    @Override
-    public BinaryDocValues getBinary(FieldInfo field) {
-        try {
-            List<BinaryDocValuesSub> subs = new ArrayList<>(this.mergeState.docValuesProducers.length);
-            for (int i = 0; i < this.mergeState.docValuesProducers.length; i++) {
-                DocValuesProducer docValuesProducer = mergeState.docValuesProducers[i];
-                if (docValuesProducer != null) {
-                    BinaryDocValues values = docValuesProducer.getBinary(field);
-                    if (values != null) {
-                        subs.add(new BinaryDocValuesSub(mergeState.docMaps[i], values));
-                    }
-                }
-            }
-            return new KNN84BinaryDocValues(DocIDMerger.of(subs, mergeState.needsIndexSort));
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-}
